crm_backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.2
django-environ==0.11.2
django-redis==5.4.0
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
redis==5.0.7
psycopg2-binary==2.9.9
Pillow==10.4.0
panoramisk==1.2
gunicorn==22.0.0
uvicorn[standard]==0.30.1
whitenoise==6.7.0
drf-spectacular==0.27.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
